added authrizations.....

+++ src/main/java/com/chatApplication/Controller/ChatController.java
@@ -3,7 +3,6 @@ package com.chatApplication.Controller;
 import com.chatApplication.Model.Chat;
 import com.chatApplication.Model.User;
 import com.chatApplication.Services.ChatService;
-import com.chatApplication.Services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,16 +14,32 @@ import java.util.NoSuchElementException;
 @RestController
 @RequestMapping("/chat")
 public class ChatController {
+    private static final String defaultAuthValue = "chat12345";
 
+    public Boolean authorize(String authValue) {
+        return defaultAuthValue.equals(authValue);
+    }
 
     @Autowired
     ChatService chatService;
 
 
     @GetMapping(" ")
-    public List<Chat> chatList() {
-        return chatService.listAllChat();
+    public ResponseEntity<Object> chatList(@RequestHeader ("authorization") String authValue ) {
+
+        if (authorize(authValue)) {
+            List<Chat> chatList = chatService.listAllChat();
+            // check if database is empty
+            if (chatList.isEmpty()) {
+                return new ResponseEntity<>("No data available", HttpStatus.NOT_FOUND);
+            } else {
+                return new ResponseEntity<>(chatList, HttpStatus.OK);
+            }
+
+        } else return new ResponseEntity<>("Not authorize", HttpStatus.UNAUTHORIZED);
     }
+
+
 
     @GetMapping("/get")
     public ResponseEntity<Chat> get(@RequestParam ("question") Long id) {
@@ -38,24 +53,38 @@ public class ChatController {
 
 
     @PostMapping("/add")
-    public void add(@RequestBody Chat chat) {
-        chatService.save(chat);
+    public ResponseEntity<String> addUser(@RequestHeader("Authorization") String authValue, @RequestBody Chat chat) {
+        //check authorization
+        if (authorize(authValue)) {
+            chatService.save(chat);
+            return new ResponseEntity<>("chat added successfully", HttpStatus.CREATED);
+        } else return new ResponseEntity<>(" not authorize ", HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/update")
-    public ResponseEntity<?> update(@RequestBody Chat chat) {
-
+    public ResponseEntity<Object> update(@RequestHeader("Authorization") String authValue, @RequestBody Chat chat) {
+        if (authorize(authValue)) {
         try {
             chatService.save(chat);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return new ResponseEntity<>("updated successfully",HttpStatus.OK);
         } catch (NoSuchElementException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+            return new ResponseEntity<>("User not found ",HttpStatus.NOT_FOUND);
+        }}
+        else return new ResponseEntity<>(" not authorize ", HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) {
-        chatService.delete(id);
+    public ResponseEntity<String> delete(@RequestHeader("Authorization") String authValue, @PathVariable Long id) {
+        if (authorize(authValue)) {
+            try {
+                chatService.delete(id);
+                return new ResponseEntity<>("User deleted successfully ",HttpStatus.OK);
+            }catch (NoSuchElementException e){
+                return new ResponseEntity<>("User not found ", HttpStatus.NOT_FOUND);
+
+            }
+
+        } else return new ResponseEntity<>(" not authorize ", HttpStatus.UNAUTHORIZED);
     }
 
 }
